Organize formatter namespace
