+++ vint/linting/formatter/json_formatter.py
